LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/template-shadowrootmode-link-stylesheet.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <div><template shadowrootmode="closed"><link rel=stylesheet href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=4bdd6438-73b9-4b3d-a8c7-58b5e5aace62&amp;encodingcheck=&Gbreve;"></template></div>


FAIL Speculative parsing, document.write(): template-shadowrootmode-link-stylesheet Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
